Move similarity scorer to analysers dir
